добавлен последовательный вариант сортировки слиянием

+++ pechishev_vladimir/lw5/lw5/stdafx.h
@@ -16,7 +16,7 @@
 #include <memory>
 #include <fstream>
 
-using Array = std::vector<float>;
+using Array = std::vector<int>;
 
 
 // TODO: Установите здесь ссылки на дополнительные заголовки, требующиеся для программы
